backend/pyproject.toml
[project]
name = "hypomnema"
version = "0.1.0"
description = "Automated Ontological Synthesizer"
license = "AGPL-3.0-or-later"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "aiosqlite>=0.20,<1",
    "sqlite-vec>=0.1.6,<1",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "anthropic>=0.42,<1",
    "google-genai>=1.5,<2",
    "apscheduler>=3.11,<4",
    "python-multipart>=0.0.18,<1",
    "pypdf>=5.0,<6",
    "python-docx>=1.1,<2",
    "feedparser>=6.0,<7",
    "youtube-transcript-api>=1.0,<2",
    "umap-learn>=0.5,<1",
    "scikit-learn>=1.6,<2",
    "httpx>=0.28,<1",
    "trafilatura>=2.0,<3",
    "cryptography>=44,<45",
    "openai>=1.60,<2",
    "platformdirs>=4,<5",
    "opendataloader-pdf>=2.0,<3",
]

[project.optional-dependencies]
local-embeddings = ["sentence-transformers>=3.4,<4"]

[dependency-groups]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.25,<1",
    "mypy>=1.14,<2",
    "ruff>=0.9,<1",
]
desktop = ["pyinstaller>=6,<7"]

[project.scripts]
hypomnema = "hypomnema.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hypomnema"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["sqlite_vec"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["sentence_transformers"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["google.genai", "google.genai.types"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["openai"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["cryptography.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["feedparser"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["trafilatura"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["apscheduler.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["umap"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["scipy.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["sklearn.*"]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_uninitialized = true

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B017"]
"src/hypomnema/api/**" = ["TC001"]
"src/hypomnema/main.py" = ["TC003"]
